requirements.txt
pandas>=2.1.1
PyYAML>=6.0.1
boto3>=1.28.0
pytest>=7.4.0
pytest-cov>=4.1.0
